requirements.txt
Django==3.1.2
djangorestframework>=3.11.0,<3.12.0
django-rest-swagger>=2.2.0,<2.3.0
pytest-django
Pillow
django-taggit
django-taggit-serializer
django-cleanup
django-imagekit
